geoguessr_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
google-auth-oauthlib==1.2.1
google-auth==2.37.0
python-dotenv==1.0.1
pika==1.3.2
pydantic==2.10.6
pytest==8.3.4
websockets==14.1
